# Minor optimizations in middleware