@sigmavirus24 2016-02-18
Each of these optimizations should be separate commits if not separate pull requests that way they can be reverted independently if necessary.